error handling and QoL updates

+++ src/services/Service-Helpers/updateWeeklyLadder/determinateWeeklyWinners.js
@@ -4,21 +4,30 @@ import WeeklyWinnersRecord from "../../../Models/WeeklyWinnersRecord.js";
 import { buildSnapshots } from "./buildSnapshots.js";
 import formatWeeklyData from "./formatWeeklyData.js";
 /**
- * @returns {Promise<void>}
+ * @returns {Promise<boolean>}
  */
 export default async function determinateWeeklyWinners() {
-    const guildCharList = await Char.find({ guildMember: true });
-    const data = await formatWeeklyData(guildCharList);
 
-    const exist = await charWeeklySnapshot.find().lean();
-    if (exist.length !== 0) {
-        await charWeeklySnapshot.collection.drop().catch((err) => {
-            if (err.code !== 26) throw err; // code 26 = "NamespaceNotFound"
-        });
+    try {
+        const guildCharList = await Char.find({ guildMember: true });
+        const data = await formatWeeklyData(guildCharList);
+        if (!data) return false;
+
+        const exist = await charWeeklySnapshot.find().lean();
+        if (exist.length !== 0) {
+            await charWeeklySnapshot.collection.drop().catch((err) => {
+                if (err.code !== 26) throw err; // code 26 = "NamespaceNotFound"
+            });
+        }
+    
+        await WeeklyWinnersRecord.create(data);
+    
+        const newWeeklySnapshotDocs = await buildSnapshots()
+        return true
+        
+    } catch (error) {
+        console.error(error);
+        return false
     }
-
-    await WeeklyWinnersRecord.create(data);
-
-    const newWeeklySnapshotDocs = await buildSnapshots()
     
 }
